src/requirements.txt
google_api_python_client==2.111.0
google_auth_oauthlib==1.2.0
langchain==0.0.352
langchain_core==0.1.3
protobuf==4.25.1
pydantic==1.10.13
python-dotenv==1.0.0
Requests==2.31.0
uagents==0.9.2
openai==1.6.1
